Format buffer contents instead of from disk (#37)
* Format buffer contents instead of from disk

Fixes #36

Pipe the current contents of the buffer to stdin on the dartfmt call
rather than have it read the file from disk.

Also improve error parsing:
- Strip the first 2 lines of the stderr output
- Prependg the buffer file path to every line, since the output now
  contains 'stdin' instead of an actual file path
- Use a multiline errorformat string so that it drops the code and
  pointer lines leaving only the error message

* Save and restore window view when rewriting buffer

When overwriting the entire buffer the cursor will normally jump to the
top of the file. Saving and restoring window view keeps cursor position
and scrolling.