Add an FAQ entry for using Dart analysis server (#81)
Closes #79

Removes mention of `dart_language_server` since the analysis server is
the supported path expect to use and promote going forward. Adds an FAQ
entry which describes the command that needs to be run to start the
analysis server since it's not as simple as a single command already on
your `$PATH`. Also points out the option of using a preconfigured plugin
which handles locating the right command for you.