Remove support for dart2js and dartanalyzer (#91)
Closes #89

The dartanalyzer support has been broken by a changed errorformat and
has not been used enough for anyone to complain. The dart2js support has
had little usage and is only usable from an entrypoint to start with so
most workflows likely use a different approach.

Leave an error message pointing to an issue with discussion about the
removal so that if users hit this they can comment and we can re-open.
In the long term we can remove these commands entirely.