Allow $ character in type names (#96)
Add `$` as a keyword character since it is allowed in Dart identifiers.
This fixes the behavior of  the `w` word motion as well as allows syntax
groups to match across the character.

Refactor the `dartTypeName` regex. Use an optional `_` at the beginning,
rather than repeating the pattern with and without `_`. Use the `\u`
character class instead of the group `[A-Z]`. Replace the `\w` character
class with a group that includes the same characters and adds `$`.