Scrape SDK version to pick format command
If the `dart` command exists, parse the output of `dart --version` and
for SDK version `2.14` and higher use `dart format`. On older SDKs
fallback to the existing behavior.

There will be some `2.14.0-dev` SDKs which are not compatible with this
check, but they aren't likely be to in use anymore for local
development.