Improve handling of dart format stderr output
Previously the `stdout` and `stderr` from the format execution were
merged and there was special handling for 1 specific warning message
followed by a standard 2 lines of boilerplate.

Now `stdout` an `stderr` are read separately. `stderr` is ignored unless
there is no output on `stdout`. This filters any warnings that didn't
prevent formatting, including the old hardcoded warning as well as a new
one around resolving `analysis_options.yaml` and any future warnings
that may come up.

When there is no `stdout` content treat the `stderr` as using the
diagnostic format. Ignore any lines that don't match the error format.
I'm not currently aware of any output here that we _wouldn't_ want to
hide, but if error output becomes important for the user to see we may
need to revisit this design.

- Migrate from `systemlist` to `job_start` which allows separating the
  output streams.
- Pull the result handling out into a separate function so it can be the
  callback for when the output has been read.